LUD-716 For the "no MDM" case set all facts to defaults
Some facts weren't being included by the discovery script in the case
where the ScaleIO gateway had not yet been configured with
MDMs. Because this data is just merged into the existing fact data in
puppetdb it was left there and shown in the UI in the case where a
ScaleIO gateway that previously had inventory was reset to the empty
state.