VXFM-5293 VxFM does not configure LIA password
RPM installation for LIA was not getting invoked leading to LIA RPM getting installed as default RPM without any password.
Also the password value needs to be encryped in the manifest file for security reason. Included asm_decrypt as the provided which will get the plain text password corresponding to the tokenized password included in the resource invocation